Addition/Subtraction Labeling
Here is a picture of the project with everything labeled. A binary number is created by flipping switches from 0 to 1 in the respective digit places. There are two switch blocks so two numbers can be created to be added or subtracted. The M switch switches from addition to subtraction. When it is a 0, the project adds, and when it is a 1, the project subtracts. The led lights then light up to display the correct answer with the leftmost light the overflow light and the second most left being the most significant bit.